+++ src/components/DownloadComponents/DownloadCard/DownloadCard.tsx
@@ -6,7 +6,6 @@ import BoxBorders from "../../SpriteSheet/BrushSvgs/BoxBorders";
 import P from "../../Typography/P";
 import Flex from "../../Flex";
 import Checkbox from "../../Checkbox";
-import Box from "../../Box";
 import GraphicCircleIcon from "../../Icon/GraphicCircleIcon";
 import { IconName } from "../../Icon";
 import type { DownloadResourceType } from "../downloads.types";
@@ -16,6 +15,7 @@ export type DownloadCardProps = CheckboxProps & {
   label: string;
   resourceType: DownloadResourceType;
   extension: string;
+  hasError?: boolean;
 };
 
 type DownloadCardLabelProps = DownloadCardProps & {
@@ -82,12 +82,20 @@ const DownloadCardLabel: FC<DownloadCardLabelProps> = ({
 );
 
 const DownloadCard: FC<DownloadCardProps> = (props) => {
-  const { checked = false, onChange, id, name, label, onBlur } = props;
+  const {
+    checked = false,
+    onChange,
+    id,
+    name,
+    label,
+    onBlur,
+    hasError = false,
+  } = props;
 
   const { hoverProps, isHovered } = useHover({});
 
   return (
-    <Box $maxWidth={200} {...hoverProps}>
+    <Flex $maxWidth={200} {...hoverProps}>
       <Checkbox
         id={id}
         name={name}
@@ -96,10 +104,11 @@ const DownloadCard: FC<DownloadCardProps> = (props) => {
         variant={"cardCheckbox"}
         ariaLabel={label}
         onBlur={onBlur}
+        hasError={hasError}
       >
         <DownloadCardLabel isHovered={isHovered} {...props} />
       </Checkbox>
-    </Box>
+    </Flex>
   );
 };